requirements.txt
Django==1.9
django-bootstrap-form==3.2
django-taggit==0.18.0
